third_party/blink/web_tests/fast/forms/number/number-validity-typemismatch-expected.txt

This test aims to check for typeMismatch flag with type=number input fields

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS "0" is a valid number.
PASS "10" is a valid number.
PASS "01" is a valid number.
PASS "0.2" is a valid number.
PASS ".2" is a valid number.
PASS "-0" is a valid number.
PASS "-1.2" is a valid number.
PASS "1.2E10" is a valid number.
PASS "1.2E-10" is a valid number.
PASS "1.2E+10" is a valid number.
PASS "12345678901234567890123456789012345678901234567890" is a valid number.
PASS "0.12345678901234567890123456789012345678901234567890" is a valid number.
PASS "abc" was sanitized to "".
PASS "0xff" was sanitized to "".
PASS "+1" was sanitized to "".
PASS " 10" was sanitized to "".
PASS "10 " was sanitized to "".
PASS "0." was sanitized to "".
PASS "1." was sanitized to "".
PASS "1,2" was sanitized to "".
PASS "1E" was sanitized to "".
PASS "NaN" was sanitized to "".
PASS "nan" was sanitized to "".
PASS "Inf" was sanitized to "".
PASS "inf" was sanitized to "".
PASS "Infinity" was sanitized to "".
PASS "infinity" was sanitized to "".
PASS "" is a valid number.
PASS "1.2E65535" was sanitized to "".
PASS "1.2e10" was sanitized to "1.2e10".
PASS "invalid" is a valid number when disabled.
PASS successfullyParsed is true

TEST COMPLETE
